// src/app/modules/recipes/models/edit-recipe-form.model.ts
export interface EditRecipeForm {
    id:number;
    name: string;
    description: string;
    code: string;
    price: number;
    cost: number;
    headquarter_id: number;
    course_id: number;
    term_id: number;
    detail: { supplyId: number, quantity: number, measuredUnitId: number, }[];
}